showing all activity entries in the day view

+++ app/src/main/java/com/example/selfproductivityapp/database/ActivitiesDay.kt
@@ -7,7 +7,7 @@ import androidx.room.PrimaryKey
 @Entity(tableName = "daily_activities_table")
 data class ActivitiesDay (
     @PrimaryKey(autoGenerate = true)
-    var entryId: Int = 0,
+    var entryId: Long = 0L,
 
     @ColumnInfo(name = "start_time_milli")
     var startTimeMilli: Long = 0L,
